+++ src/game/nlcore/common/game.ts
@@ -2,6 +2,7 @@ import {Game} from "@core/game";
 import {GameState} from "@player/gameState";
 import {Storable, Namespace} from "../elements/persistent/storable";
 import {LiveGame} from "@core/game/liveGame";
+import {Preference} from "@core/game/preference";
 import type {SavedGame} from "@core/gameTypes";
 
 export {
@@ -10,6 +11,7 @@ export {
     Game,
     Storable,
     Namespace,
+    Preference,
 };
 export type {
     SavedGame,